updates to the structure of the dataloaders. pass to sherin to complete work
